requirements.txt
opencv-python==4.13.0.90
ollama==0.6.1
pytesseract==0.3.13
pandas==2.3.3
pillow==12.1.0
flask==3.1.2
openpyxl==3.1.5
google-cloud-vision==3.12.0
numpy==2.2.6
requests==2.32.5
werkzeug==3.1.5
jinja2==3.1.6
pydantic==2.12.5
